requirements.txt
tensorflow>=2.16.1
fastapi
uvicorn
python-multipart
jinja2
pillow
numpy
pandas
plotly
scikit-learn
matplotlib
seaborn
requests
